requirements_nuevos.txt
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
PyYAML==6.0.3
uritemplate==4.2.0